PCLVisualizer: save and restore camera information
Press ctrl + s to save camera parameters and ctrl + r to restore the
saved camera parameters.

If a camera file is specified with "-cam" option or pcd file(s) is(are)
available in the command line, the camera parameters will be written to
the corresponding file and can be restored later from that file.
In this case, PCLVisualizer will restore camera parameters automatically
from the corresponding file when the pragram restarts with the same
command line input.