prefer qhull_p shared libraries over libqhull
upstream added libqhull_p.so in version 2012 which is compiled
_with_ -Dqh_QHpointer and changed libqhull.so to use -Dqh_QHpointer=0.

Only debian/ubuntu doesn't acknowledge this and still provides
libqhull.so with -Dqh_QHpointer=1.

Before this change PCL+QHULL_USE_STATIC=0 compiled with -Dqh_QHpointer,
but always linked libqhull.so. As explained this is the wrong choice on
qhull 2012 with non-debian systems.

Closes #410