Update Travis scripts to migrate to container-based infrastructure
Travis is currently migrating from legacy to container-based
infrastructure. Benefits include quick build starting time, caching,
more available resources, and no notorious 50 minutes limit.  One of the
difficulties though is that on container-based infrastructure we have no
sudo rights, which means that the additional software that we need may
only be installed either from white-listed repositories, or from source.
Not all of PCL dependencies are available in the white-listed PPAs, for
example there is no VTK and FLANN.

Thanks to caching, installation of dependencies from source is not a big
deal.  Effectively, it even reduces build time, because once the
software is built, installed, and cached, Travis doesn't need to spend
time to download and install it from scratch on each run.

The list of dependencies that the updated script installs:

 * From white-listed repositories:
     - CMake
     - Boost 1.55
     - Eigen 3
     - GTest
     - Latex packages for Doxygen
     - libusb 1.0

 * From source:
     - VTK 5.10.1
     - FLANN 1.8.4
     - QHull 2012.1t
     - Doxygen 1.8.9.1

According to my tests, compilation task succeeds in 100% cases. Test
task manages to build everything, however one of the unit tests fails.
We will need to either fix it or disable for the time being.