Fix: invalid deallocation of arrays used by VTK
Within pcl/visualization, the method SetArray() from the vtkDataArray is
called from several locations in the source code. If not specified, the
vtkDataArray uses free() to deallocate the given memory. But in most
cases (in pcl/visualization) the allocation of arrays is done by new[].
Therefore, VTK_DATA_ARRAY_DELETE was appended as fourth parameter of the
SetArray() method to force the use of delete[] instead.