Fix occasional seg fault with KLDAdaptiveParticleFilterOMPTracker
During the resampling step of the KLDAdaptiveParticleFilterTracker, a
situation occasionally occurs where the number of particles generated
exceeds the maximum_particle_number_. This results in a vector being
accessed past its end, leading to a segfault. This happens when the
variable k in the resampling step is less than 2. To fix this, the
condition on the while loop has been slightly modified.