Remove depreciation warning from sample_size_pairs
This warning floods build output because the variable is used below in
SAC_SAMPLE_SIZE. Even though the latter is never used throughout PCL
code base, inclusion of the header is sufficient to trigger the warning.

This commit does not touch the depreciation warning of SAC_SAMPLE_SIZE,
so anyone who happens to use the old model size determination system
will still be notified.