Fix error CMP0020
PCL that include Qt5 can't generate project,  because following errors
occurs in CMake outputs on Windows.

"Policy CMP0020 is not set:
Automatically link Qt executables to qtmain target on Windows.
Run "cmake --help-policy CMP0020" for policy details.
Use the cmake_policy command to set the policy and suppress this
warning.
This warning is for project developers.
Use -Wno-dev to suppress it."

Fix this issue by setting the CMake policy.