Fix a bug in FPCSInitialAlignment
`target_indices_` is a pointer. Sample consensus model does not have a
constructor that would accept a pointer to indices vector. However, it
has a constructor where second parameter is a boolean. Therefore an
implicit cast happens here and instead of setting indices we actually
request using random seed.  This commit sets indices in a separate call.
Another option would be to pass by reference to value to the
constructor, however this would involve copying the vector instead of
simply sharing.