Fixes PLY parsing of char and uchar scalars
WIP of #1381. int8 and uint8 types were being evaluated as char and
unsigned char, their actual original type, so whenever a lexical_cast
was performed from a multi-digit string to an (unsigned) char, a
bad_lexical_cast exception was thrown and the value set to a quiet
NaN. Functionally-wise, this PR will make color parsing work.