Explicitly set VCPKG_ROOT environment variable
At the moment this variable is not set in the Azure agent images. This
was already fixed in Azure image generation repository, but it will take
up to three weeks until a new release happens. This commit temporary
fixes the issue and should be reverted once new images are deployed.

See: Microsoft/azure-pipelines-image-generation#769