Update uniform_sampling.h
Please check filters/impl/uniform_sampling.hpp where the removed_indices are set. The implementation does smell a bit but the comments say that all the points present will be approximated (i.e., *downsampled*) with the closest point to the center of the voxel.

Maybe its also worth to mention that it calculates point indices and is not producing a new cloud like the VoxelGrid Class.