Octree: grow in positive direction instead of negative Fixes https://github.com/PointCloudLibrary/pcl/issues/5637 The octree must grow if a point lies outside in (at least) one axis. For each axis, the bounding box may be violated in the lower bound, or upper bound, or neither. Currently, in the last case, the tree will grow in the negative direction. With these changes it will instead grow in the positive direction. This is better because the points will be closer to min_*_, which will lead to smaller errors when computing the key.
It is necessary to adapt the GFPFH test because GFPFH uses an octree and the feature is influenced by which octree leafs are occupied exactly.
Previously, the test cloud resulted in an octree with bounding box (-108;20)(-104;24)(-72;56), with these changes the bounding box is (-12;20)(-12;20)(-12;20), which makes more sense.
I am also not sure how much sense this test makes because the test assignes the class labels more or less randomly, while the paper says they should represent a "geometric primitive" (plane, sphere, ...).